bot/requirements.txt
aiogram==3.19.0
aiohttp==3.11.14
python-dotenv==1.0.0
pytz==2023.3.post1
python-dateutil==2.8.2